tests/kano_i18n/test_init.py
import base64
import os
import pytest
from gettext import GNUTranslations, NullTranslations
from test.test_support import EnvironmentVarGuard

from kano_i18n.init import register_domain, install, get_current_translation


# From cpython test_gettext.py
GNU_MO_DATA = '''\
3hIElQAAAAAGAAAAHAAAAEwAAAALAAAAfAAAAAAAAACoAAAAFQAAAKkAAAAjAAAAvwAAAKEAAADj
AAAABwAAAIUBAAALAAAAjQEAAEUBAACZAQAAFgAAAN8CAAAeAAAA9gIAAKEAAAAVAwAABQAAALcD
AAAJAAAAvQMAAAEAAAADAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAABQAAAAYAAAACAAAAAFJh
eW1vbmQgTHV4dXJ5IFlhY2gtdABUaGVyZSBpcyAlcyBmaWxlAFRoZXJlIGFyZSAlcyBmaWxlcwBU
aGlzIG1vZHVsZSBwcm92aWRlcyBpbnRlcm5hdGlvbmFsaXphdGlvbiBhbmQgbG9jYWxpemF0aW9u
CnN1cHBvcnQgZm9yIHlvdXIgUHl0aG9uIHByb2dyYW1zIGJ5IHByb3ZpZGluZyBhbiBpbnRlcmZh
Y2UgdG8gdGhlIEdOVQpnZXR0ZXh0IG1lc3NhZ2UgY2F0YWxvZyBsaWJyYXJ5LgBtdWxsdXNrAG51
ZGdlIG51ZGdlAFByb2plY3QtSWQtVmVyc2lvbjogMi4wClBPLVJldmlzaW9uLURhdGU6IDIwMDAt
MDgtMjkgMTI6MTktMDQ6MDAKTGFzdC1UcmFuc2xhdG9yOiBKLiBEYXZpZCBJYsOhw7FleiA8ai1k
YXZpZEBub29zLmZyPgpMYW5ndWFnZS1UZWFtOiBYWCA8cHl0aG9uLWRldkBweXRob24ub3JnPgpN
SU1FLVZlcnNpb246IDEuMApDb250ZW50LVR5cGU6IHRleHQvcGxhaW47IGNoYXJzZXQ9aXNvLTg4
NTktMQpDb250ZW50LVRyYW5zZmVyLUVuY29kaW5nOiBub25lCkdlbmVyYXRlZC1CeTogcHlnZXR0
ZXh0LnB5IDEuMQpQbHVyYWwtRm9ybXM6IG5wbHVyYWxzPTI7IHBsdXJhbD1uIT0xOwoAVGhyb2F0
d29iYmxlciBNYW5ncm92ZQBIYXkgJXMgZmljaGVybwBIYXkgJXMgZmljaGVyb3MAR3V2ZiB6YnFo
eXIgY2ViaXZxcmYgdmFncmVhbmd2YmFueXZtbmd2YmEgbmFxIHlicG55dm1uZ3ZiYQpmaGNjYmVn
IHNiZSBsYmhlIENsZ3ViYSBjZWJ0ZW56ZiBvbCBjZWJpdnF2YXQgbmEgdmFncmVzbnByIGdiIGd1
ciBUQUgKdHJnZ3JrZyB6cmZmbnRyIHBuZ255YnQgeXZvZW5lbC4AYmFjb24Ad2luayB3aW5rAA==
'''


@pytest.yield_fixture
def test_env():
    with EnvironmentVarGuard() as env:
        yield env


def test_register_domain_fallback(tmpdir, test_env):
    test_env['LANG'] = 'en_GB'
    locale_dir = tmpdir.ensure('en_GB', 'LC_MESSAGES', dir=True)

    # setup top level domain (has no translations)
    install('app-domain', locale_dir=tmpdir.strpath)

    domain1_mo = os.path.join(locale_dir.strpath, 'domain-1.mo')
    with open(domain1_mo, 'wb') as fp:
        fp.write(base64.decodestring(GNU_MO_DATA))

    # register a new domain (with translations)
    register_domain('domain-1', locale_dir=tmpdir.strpath)

    # inspect domain setup
    current_translation = get_current_translation()
    assert current_translation is not None
    assert isinstance(current_translation, NullTranslations)
    assert current_translation._fallback is not None
    assert current_translation._fallback._fallback is None
    assert isinstance(current_translation._fallback, GNUTranslations)

    # No translation defined
    assert _('Hi') == 'Hi'
    assert _('nudge nudge') == 'wink wink'
